arranged modules and dependency files to separate folders
