+++ EmailSubject.txt
@@ -1,1 +1,1 @@
-Pictorial Reminder for February 14
+(CORRECTION:) Pictorial Reminder for February 15
