--- merlin.json
@@ -1,12 +1,0 @@
-{
-  "type": "service_account",
-  "project_id": "merlin-333803",
-  "private_key_id": "904172883465b2d77d9c9a60278f5bed0ca76edd",
-  "private_key": "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDI0c5LxIrhr8Av\nvhTHRzPcBG1w/iam7aPsnT5Ob03l6MnpvfMyrF3SeI+hwcMabZtUhsSbGBGmuSBY\nhAy+INM93fB7Cg1lv8n+njU/5Oqt7uVMYAuvClMk//sb5Gmoo86esjm9JOiW1Thc\nFQk47Yk7cmW8j1GxQSn3bZISMOjkaglrkj6lG6tLjVDKGePPt7rAVyU8rpIAa9Ap\naR/aQuhzES8KjH2fMDCduS4SUW8zwlk8ASldutGRgZ4LLK1VeVgk5/6vWXQl/YZr\ntm7VvFzVCyPT+WYNkNz3N4QGzEgD43cML0jiSOCUoboemx4277JPvIM9BJKHzs6M\nLvwCoP/DAgMBAAECggEAFR321wHSZtwP1hVjYvcqTWMLzRt0ng4UND/uutm6c8gw\nX0txpOz6KrUxyV0yMne/IzrOts5noMQL0jMEkgOqovFRJecPNJlf2M3zJIZiZSrf\nYtIguGXaC2R9l9fDgo/DHO3541lbueyu8gPmRd24lg84AoWBniWPj6yudQo+mzIZ\nnt825sz2J4WJf9l+ctsnui1Ple5r2Z4l+2+AOzKwt3bfaT+nIUtuuzSKxwMm42kt\n47iPcOaCrFkRKvnVOJbAcOw9pRiDblpEwmfvrHwkoMYgYG/SJNd/HC/GJe6ctPfO\n7SLOpphYlG95cEYP3dZxr5ELyY2k7XWoEuPzGs6hpQKBgQDjvHJdlPbWIwZxNAPS\nwniS63sujSnjg3IbhFBo4nrcAi7iVIYjlmDye/r9uLqTgdTlh5iqUWsUweFFqLEu\ntFMFGhxYYiQ6E5obp2XjrmVYYK1mDEF/qS3zY7caovjUgHlej/C/jsBpdPpV5LEh\nd53GVGG2tz5vKEBi9oKUkTkH1wKBgQDhvize2iyPmczW4SRzOASSQ7SA5sGErXPc\nS+NdeQ6AXySTLL6zAQGdBOzT7Pk1iIGCdLPcqRfUCIbmGYZ23l6LqqbaUFKI9Xr7\nmOtexAPb8tfNLOGmgu9LIvMqYXBbGsnruZpRCIhwtI5zq5We6s7PUofTprBe4RXl\nF2jCWeuZ9QKBgFi6jVyEs39O+i2U9IBjIJMtiDrk6RMpnYW3EZUr2G5rs1SfVfv7\nlR4ky++az6a1CPu8zvBc9HmL6Hc5eH8X5tiwCact5atzA7IOkJYa1oWOdUCzObnB\ncpPHoAmSwZwZyi/bqRMZTUJJP1fD2HePoTsqmT+VOZAZgKyOGuZDhpLVAoGBAJh2\n4ElXY3dNchrDDuI+uF+tGLFiqNsaZ0EnhBVymIe9Oq+oLdgqWtOZKvbre1lbq9TX\nS09CaVSMuceY/8cy4vpDtNOIImZcHcbl/joP+2dHOddac8EJd+LwZPl/u+ncSvpJ\nbd9C+xvb5sZa4f/mKO2/mPsq/r3mGeqTh+6T2tstAoGAeww1CtKZlzOA7CCq2AkJ\niZlhffyVxuPCtd8NVpoJthkkFRHTR/Dn2co6X186lsLOsMF2pGxY4qEab8j8Zh46\n0+GJREnCBCz9hBWmPwO9KZS4tE0gmijeP7O13/8vJnz3+GQxfCjfCYH8bZkQMcq6\n3actSJiCevyiuvrzicC8u/E=\n-----END PRIVATE KEY-----\n",
-  "client_email": "merlin@merlin-333803.iam.gserviceaccount.com",
-  "client_id": "102365728640771610561",
-  "auth_uri": "https://accounts.google.com/o/oauth2/auth",
-  "token_uri": "https://oauth2.googleapis.com/token",
-  "auth_provider_x509_cert_url": "https://www.googleapis.com/oauth2/v1/certs",
-  "client_x509_cert_url": "https://www.googleapis.com/robot/v1/metadata/x509/merlin%40merlin-333803.iam.gserviceaccount.com"
-}
