+++ modules/Search.py
@@ -2,13 +2,13 @@ import PyQt5
 from PyQt5.QtWidgets import QMessageBox
 
 
-import Sheets
-import Subscriber
+import modules.Sheets as sheets
+import modules.Subscriber as sub
 
 
 def SearchUserID(ID):
     try:
-        subscriberData = Sheets.SearchID(ID)
+        subscriberData = sheets.SearchID(ID)
         # Segments the subscriber data list into individual pieces of information that can be used to change the UI as well as the email body message
 
         email = subscriberData[1]
@@ -18,7 +18,7 @@ def SearchUserID(ID):
         time = subscriberData[14]
         phoneNumber = subscriberData[6]
         status = subscriberData[15]
-        subscriber = Subscriber.Subscriber(
+        subscriber = sub.Subscriber(
             fname, lname, email, date, time, phoneNumber, status)
         print('abot dinhi')
         return subscriber
@@ -31,7 +31,7 @@ def SearchLastName(surname):
     try:
         print('in try')
         # chops up the data inside string to specific variables
-        subscriberData = Sheets.SearchLastName(surname)
+        subscriberData = sheets.SearchLastName(surname)
         print(f'subscriber data is {subscriberData}')
         email = subscriberData[1]
         fname = subscriberData[4]
@@ -41,7 +41,7 @@ def SearchLastName(surname):
         phoneNumber = subscriberData[6]
         status = subscriberData[15]
         # instantiates a subscriber object from the subscriber class
-        subscriber = Subscriber.Subscriber(
+        subscriber = sub.Subscriber(
             fname, lname, email, date, time, phoneNumber, status)
         return subscriber
     except:
